requirements.txt
jupyterlab
ipykernel
ipywidgets
PyMuPDF
openai
numpy
python-dotenv
youtube-transcript-api
flask